refactor(actions/pick): use fetchRestful method
